ai_services/pyproject.toml
[project]
name = "ai-services"
version = "0.1.0"
description = ""
authors = [
    {name = "ZOUARIWassim",email = "wassim.zouari07@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    "transformers (>=4.56.1,<5.0.0)",
    "ipykernel (>=6.30.1,<7.0.0)",
    "torch (>=2.8.0,<3.0.0)",
    "ipywidgets (>=8.1.7,<9.0.0)",
    "google-genai (>=1.33.0,<2.0.0)",
    "fastapi (>=0.116.1,<0.117.0)",
    "google-cloud-storage (>=3.3.1,<4.0.0)",
    "jinja2 (>=3.1.6,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.prod.dependencies]
google-cloud-speech = "^2.33.0"


[tool.poetry.group.dev.dependencies]
google-cloud-speech = "^2.33.0"


[tool.poetry.group.notebook.dependencies]
jupyter = "^1.1.1"
